requirements.txt
Django
djangorestframework
azure
Celery
